# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Think Then Talk API
  description: The complete API specification for the Think Then Talk event platform.
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: Development server
...
paths:
  /auth/request-otp:
    post:
      tags:
        - Auth
      summary: Request an OTP
      description: Initiates the login process by sending a one-time password to the user's mobile number.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RequestOtpBody'
      responses:
        '200':
          description: OTP sent successfully.
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
        '400':
          description: Invalid mobile number format.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify-otp:
    post:
      tags:
        - Auth
      summary: Verify OTP
      description: Verifies the OTP and logs the user in, returning a session token and user data.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifyOtpBody'
      responses:
        '200':
          description: Login successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                    description: A session token (e.g., JWT).
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid OTP.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /admin/login:
    post:
      tags:
        - Auth
      summary: Admin Login
      description: Authenticates an admin user via email and password.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminLoginBody'
      responses:
        '200':
          description: Admin login successful.
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  user:
                    $ref: '#/components/schemas/AdminUser'
        '401':
          description: Invalid credentials.
...
  /admin/sponsors:
    get:
      tags:
        - Content Management
      summary: List All Sponsors (Admin)
      description: Retrieves all sponsors for the admin panel.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of all sponsors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Sponsor'
    post:
      tags:
        - Content Management
      summary: Create Sponsor (Admin)
      description: Adds a new sponsor.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorFormData'
      responses:
        '201':
          description: Sponsor created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
  '/admin/sponsors/{sponsorId}':
    put:
      tags:
        - Content Management
      summary: Update Sponsor (Admin)
      description: Updates an existing sponsor's details.
      security:
        - bearerAuth: []
      parameters:
        - name: sponsorId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SponsorFormData'
      responses:
        '200':
          description: Sponsor updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Sponsor'
        '404':
          description: Sponsor not found.
    delete:
      tags:
        - Content Management
      summary: Delete Sponsor (Admin)
      description: Deletes a sponsor.
      security:
        - bearerAuth: []
      parameters:
        - name: sponsorId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Sponsor deleted.
        '404':
          description: Sponsor not found.
...
  /admin/team:
    get:
      tags:
        - Content Management
      summary: List All Team Members (Admin)
      description: Retrieves all team members for the admin panel.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of all team members.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TeamMember'
    post:
      tags:
        - Content Management
      summary: Create Team Member (Admin)
      description: Adds a new team member.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberFormData'
      responses:
        '201':
          description: Team member created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
  '/admin/team/{memberId}':
    put:
      tags:
        - Content Management
      summary: Update Team Member (Admin)
      description: Updates an existing team member's details.
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamMemberFormData'
      responses:
        '200':
          description: Team member updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TeamMember'
        '404':
          description: Team member not found.
    delete:
      tags:
        - Content Management
      summary: Delete Team Member (Admin)
      description: Deletes a team member.
      security:
        - bearerAuth: []
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Team member deleted.
        '404':
          description: Team member not found.
...
components:
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
          description: A human-readable error message.
      required:
        - message
    RequestOtpBody:
      type: object
      properties:
        mobile:
          type: string
          description: The user's 11-digit mobile number.
          example: '09123456789'
      required:
        - mobile
    VerifyOtpBody:
      type: object
      properties:
        mobile:
          type: string
          example: '09123456789'
        otp:
          type: string
          description: The 6-digit one-time password.
          example: '123456'
      required:
        - mobile
        - otp
    AdminLoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
      required:
        - email
        - password
...
    User:
      type: object
      required:
        - id
        - firstNameFa
        - lastNameFa
        - mobile
      properties:
        id:
          type: string
          description: The unique identifier for the user.
        firstNameFa:
          type: string
          description: User's first name in Persian.
        lastNameFa:
          type: string
          description: User's last name in Persian.
        firstNameEn:
          type: string
        lastNameEn:
          type: string
        gender:
          type: string
          enum:
            - MALE
            - FEMALE
            - OTHER
        age:
          type: number
          format: int32
        educationLevel:
          type: string
        fieldOfStudy:
          type: string
        isEmployed:
          type: boolean
        jobTitle:
          type: string
        mobile:
          type: string
          description: The user's primary mobile number (unique).
        email:
          type: string
          format: email
        languageLevel:
          type: string
...
    AdminUser:
      type: object
      required:
        - id
        - email
        - name
        - role
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        name:
          type: string
        role:
          type: string
          enum:
            - ADMIN
            - EVENT_MANAGER
            - FINANCE
...
    Sponsor:
      type: object
      required:
        - id
        - name
        - productOrTagline
        - logoUrl
      properties:
        id:
          type: string
        name:
          type: string
        productOrTagline:
          type: string
        logoUrl:
          type: string
          format: uri
        websiteUrl:
          type: string
          format: uri
    SponsorFormData:
      type: object
      required:
        - name
        - productOrTagline
        - logoUrl
      properties:
        name:
          type: string
        productOrTagline:
          type: string
        logoUrl:
          type: string
          format: uri
        websiteUrl:
          type: string
          format: uri
    TeamMember:
      type: object
      required:
        - id
        - name
        - role
        - photoUrl
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        photoUrl:
          type: string
          format: uri
    TeamMemberFormData:
      type: object
      required:
        - name
        - role
        - photoUrl
      properties:
        name:
          type: string
        role:
          type: string
        photoUrl:
          type: string
          format: uri
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT